refactor: move client to another package (#22)
